refactor: split bus and images packages
