+++ pkg/agentcompose/webhook_queue_test.go
@@ -115,7 +115,7 @@ func testLoaderEventLoopRetriesWhenWebhookQueueFull(t *testing.T) {
 			WebhookQueueDefaultWorkers: 1,
 		},
 		configDB:     store,
-		bus:          &LoaderBus{ch: make(chan LoaderTopicEvent, 8)},
+		bus:          NewLoaderBusWithBuffer(8),
 		engine:       &QJSLoaderEngine{},
 		eventQueue:   &WebhookRunQueue{defaultWorkers: 1, running: map[string]int{}},
 		loaders:      map[string]Loader{},
@@ -208,7 +208,7 @@ func testLoaderEventLoopRetriesWhenSkipPolicyLoaderBusy(t *testing.T) {
 			WebhookQueueDefaultWorkers: 8,
 		},
 		configDB:     store,
-		bus:          &LoaderBus{ch: make(chan LoaderTopicEvent, 8)},
+		bus:          NewLoaderBusWithBuffer(8),
 		engine:       &QJSLoaderEngine{},
 		eventQueue:   &WebhookRunQueue{defaultWorkers: 8, running: map[string]int{}},
 		loaders:      map[string]Loader{},
@@ -296,7 +296,7 @@ func testLoaderEventLoopDedupesWebhookTargetsByLoader(t *testing.T) {
 			WebhookQueueDefaultWorkers: 8,
 		},
 		configDB:     store,
-		bus:          &LoaderBus{ch: make(chan LoaderTopicEvent, 8)},
+		bus:          NewLoaderBusWithBuffer(8),
 		engine:       &QJSLoaderEngine{},
 		eventQueue:   &WebhookRunQueue{defaultWorkers: 8, running: map[string]int{}},
 		loaders:      map[string]Loader{},
@@ -386,7 +386,7 @@ func testLoaderEventLoopRunsAllWebhookTargetLoaders(t *testing.T) {
 			WebhookQueueDefaultWorkers: 8,
 		},
 		configDB:     store,
-		bus:          &LoaderBus{ch: make(chan LoaderTopicEvent, 8)},
+		bus:          NewLoaderBusWithBuffer(8),
 		engine:       &QJSLoaderEngine{},
 		eventQueue:   &WebhookRunQueue{defaultWorkers: 8, running: map[string]int{}},
 		loaders:      map[string]Loader{},
