+++ pkg/images/image_auto_backend_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package images
 
 import (
 	"context"
