+++ pkg/images/image_oci.go
@@ -1,4 +1,4 @@
-package agentcompose
+package images
 
 import (
 	"slices"
