+++ pkg/images/image_service_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package images
 
 import (
 	"context"
